Adaptación del modelo, correcciones

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/ptr/presencia/PtrSeccionPresenciasGenericType.java
@@ -19,6 +19,6 @@ public class PtrSeccionPresenciasGenericType implements Serializable {
 
     private Integer seccion;
 
-    private Long minutos;
+    private Integer minutos;
 
 }
